Technical part/Python/src/requirements.txt
customtkinter
opencv-python
pillow
numpy